Use wp_terms_checklist() in front-end meta box.
Use the WordPress admin function `wp_terms_checklist()` to generate the checklist used in the front-end editing metabox. Then, hierarchical terms will be shown hierarchically, which is nice.
This necessitates a change to the function used to save terms, becuase WordPress is kind of weird in the way it outputs terms via thei function.